// client/src/redux/types.ts
export interface Lesson {
  lesson: any;
  loading: boolean;
  deleting: boolean;
  updating: boolean;
  deletingError: any;
  updatingError: any;
  posting: boolean;
  postingError: any;
  lessonModel: boolean;
  editLessonModel: boolean;
  editLessonModelId: string | null;
  notFound: boolean;
  loadingError: boolean;
}